teste

+++ Java - primeiros passos/src/TestaConversao.java
@@ -2,7 +2,7 @@ public class TestaConversao{
 
     public static void main(String[] args){
 
-        float pontoFlutuante = 3.14f;
+        float pontoFlutuante = (float)3.14;
 
         double salario = 1270.50;
         int valor = (int) salario;
